requirements.txt
absl-py==2.2.2
accelerate 
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.6.0
anyio==4.7.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
Bottleneck==1.4.2
certifi==2025.4.26
charset-normalizer==3.4.1
distro==1.9.0
efficientnet-pytorch
exceptiongroup==1.2.0
flatbuffers==25.2.10
frozenlist==1.6.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
importlib_metadata==8.7.0
jiter==0.6.1
jpype1==1.5.2
keras==3.9.2
koeda==0.0.4
konlpy==0.6.0
libclang==18.1.1
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.4.3
namex==0.0.9
numexpr==2.10.1
numpy==2.0.1
oauthlib==3.2.2
openai==1.60.1
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pip==25.0
propcache==0.3.1
protobuf==5.29.4
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
pyyaml
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
sentencepiece
setuptools==75.8.0
scikit-learn
six==1.17.0
sniffio==1.3.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tf-keras
torch
tqdm==4.67.1
transformers
tweepy==3.10.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0




